Fix segfault in sample.c due to analigned access caused by strcpy
The strcpy calls in sample.c caused segmentation fault on kernels with CONFIG_FORTIFY_SOURCE
enabled. This occurs because strcpy is compiled as a memcpy (preceded
by a strlen), which violates memory alignment constraints and lead to a
crash.

To resolve the issue, all strcpy calls in sample.c were replaced with a
byte-by-byte copy, ensuring that no unaligned access occurs.

Tested on both kernels with and without CONFIG_FORTIFY_SOURCE to ensure
compatibility and to confirm that the changes prevent the segfault.

Signed-off-by: Massimiliano Pellizzer <massimiliano.pellizzer@canonical.com>